healthchecker pkging checks for prefix
When the `healthchecker` package is vendored with a prefix, the golang
dependency package is also prefixed, e.g. vendoring with the prefix
`cf-networking` will result in a package named
`cf-networking-golang-1-linux`. This commit adds logic to check for that prefix
to source the golang environment variables correctly when running the
packaging script.