Update watchdog for golang 1.20.6
Starting in golang 1.20.6, additional validations on Host headers
are performed during http.NewRequest(). The current mechanism for
providing the socket file name from config to watchdog is via the
url Host value. Once we pull the socket file data from that we now
need to reset the host header to be a valid value.